Renamed configs on f4dev/docs to conform to rest of codebase.
