+++ .vscode/settings.json
@@ -9,5 +9,9 @@
 	],
 	"cmake.configureOnEdit": false,
 	"cmake.configureOnOpen": false,
-	"files.associations": {}
+	"files.associations": {
+		"hw_adc.h": "c",
+		"hw_gpios.h": "c",
+		"io_chimeraconfig_v2.h": "c"
+	}
 }
