+++ firmware/dev/f4dev/src/hw/hw_gpios.c
@@ -1,6 +1,6 @@
 #include "main.h"
 
-#include "hw_gpio_config.h"
+#include "hw_gpios.h"
 #include "hw_gpio.h"
 
 const Gpio gpio_6 = { .port = GPIO_6_GPIO_Port, .pin = GPIO_6_Pin };
